// web/src/features/project-evaluation/model/types.ts
export type PlannerFormValues = {
  email: string
  projectName: string
  occupancyType: string
  countryCode: string
  jurisdictionCode: string
  farMin: string
  farMax: string
  heightMax: string
  qualitativeMin: string
  evaluationDate: string
  aestheticGoal: string
  aestheticReferences: string
}